rest/requirements.txt
Django==1.9.8
mysqlclient==1.3.7
djangorestframework==3.4.0
markdown==2.6.6
django-filter==0.13.0
djangorestframework-jwt==1.8.0
django-cors-headers==1.1.0
behave-django==0.3.0
selenium==2.53.6
gunicorn==19.7.1
django-import-export==0.7.0
django-admin-rangefilter==0.3.1